requirements.txt
deoxys==0.0.11
diffusers==0.32.2
easydict==1.13
h5py==3.13.0
matplotlib==3.7.5
numcodecs==0.12.1
numpy==1.24.4
pandas==2.0.3
Pillow==11.2.1
pyzed==1.3.0
setuptools==44.0.0
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.0
zarr==2.16.1
